Added some basic commands and more config options available.
Now worlds can be specified not to be loaded on startup.
Now worlds are always saved asynchronously unless the server has already stopped.
Current commands are /smw <help|version|load|unload|goto>.
If you try to load a world that was serialized with a newer version than the one SMW supports, a NewerFormatException will be thrown.