Fixed a bug where worlds wouldn't get unlocked on unload. Partially implemented 1.14 support.
So, I felt that I had to try again, and I decided to took the javassist approach. Although it's a bit messy (and I don't like the idea of having to edit the startup parameters of the server just for this), I've managed to override the chunk load method. I'll deal with the saving method later. Also had to update the slime format to v5, as skylight and blocklight are not always present anymore.