Chunks are now converted to NMS when the world loads.
There's no need to convert the chunk when requested, they can be all loaded when the world is created. Moreover, if a chunk is already loaded, it's unnecessary to keep a SlimeChunk object, it's just a waste of memory. Instead, SlimeChunks can be converted to their NMS counterparts when the CustomWorldServer is created, and then they can all be deleted. When the server wants to save the world, all chunks are converted again to SlimeChunk objects, then the world is saved, and finally the SlimeChunk map is deleted again.