Add beds to the 1.11 world upgrader.
When updating to the latest SRF version, world version is deduced based on how blocks are stored on chunks, so it's always either v1 (1.8) or v4 (1.13). Because of this, some worlds might be upgraded unnecessarily, as the world version is not correct. This change adds the new bed tile entity so v5 (and older) 1.11 worlds can be loaded.