Preload nms chunks (#81)
* Chunks are now converted to NMS when the world loads.

There's no need to convert the chunk when requested, they can be all loaded when the world is created. Moreover, if a chunk is already loaded, it's unnecessary to keep a SlimeChunk object, it's just a waste of memory. Instead, SlimeChunks can be converted to their NMS counterparts when the CustomWorldServer is created, and then they can all be deleted. When the server wants to save the world, all chunks are converted again to SlimeChunk objects, then the world is saved, and finally the SlimeChunk map is deleted again.

* Chunk conversion shold be done after the save timing has started.

* Preload nms chunks on spigot 1.9.

* Preload nms chunks on spigot 1.9.4.

* Preload nms chunks on spigot 1.10.2.

* Preload nms chunks on spigot 1.11.

* Chunk is never going to be null.

* Preload nms chunks on spigot 1.12.2.

* Preload nms chunks on spigot 1.13.

* Forgot to update the save chunk method on spigot 1.13.

* Preload nms chunks on spigot 1.13.2.

* Re-add chunks to the list before saving.

* Preload nms chunks on spigot 1.14.

* Preloading spawn chunks isn't necessary, as all chunks are preloaded.